feat(queue): implement ranked match recruitment and voting system
- Added `executeRank` function to handle ranked match recruitment, including embed creation and API interaction for recruitment creation.
- Implemented `handleVote` function to manage voting for matches, including participant validation and vote confirmation.
- Created `handleVoteCancel` function to allow participants to cancel their votes.
- Developed team balancing logic in `balanceTeamsByElo` to distribute participants based on Elo ratings.
- Introduced button handling functions for joining, leaving, and managing recruitment sessions.
- Added constants and utility functions for role management and embed creation.
- Enhanced embed creation functions to support new features and display participant information effectively.
- Established types for participants and team assignments to improve type safety and clarity in the codebase.
