+++ src/commands/queue/create/index.ts
@@ -1,15 +1,16 @@
 import type { ChatInputCommandInteraction } from 'discord.js'
 import { logger } from '@/lib/logger'
 import { apiClient } from '@/utils/api-client'
-import { CAPACITY, createRankedButtons, createRankedEmbed } from './shared'
+import { CAPACITY } from '../shared/constants'
+import { createButtons, createEmbed } from '../shared/embeds'
 
-export const executeRank = async (interaction: ChatInputCommandInteraction, guildId: string) => {
+export const executeCreate = async (interaction: ChatInputCommandInteraction, guildId: string) => {
 	const description = interaction.options.getString('description')
 	const startTime = interaction.options.getString('start_time')
 	const recruitmentId = crypto.randomUUID()
 
-	const embed = createRankedEmbed([], CAPACITY, interaction.user.id, startTime, description)
-	const disabledButtons = createRankedButtons(recruitmentId, true)
+	const embed = createEmbed(false, [], CAPACITY, interaction.user.id, startTime, description)
+	const disabledButtons = createButtons(recruitmentId, true)
 
 	await interaction.reply({
 		embeds: [embed],
@@ -26,31 +27,30 @@ export const executeRank = async (interaction: ChatInputCommandInteraction, guil
 				channelId: interaction.channelId,
 				messageId: reply.id,
 				creatorId: interaction.user.id,
-				type: 'ranked',
 				anonymous: false,
 				startTime: startTime || undefined,
 			},
 		})
 
 		if (!response.ok) {
-			logger.error('ランク戦募集作成失敗:', response.status)
+			logger.error('募集作成失敗:', response.status)
 			await interaction.editReply({
-				content: 'ランク戦募集の作成に失敗しました。',
+				content: '募集の作成に失敗しました。',
 				embeds: [],
 				components: [],
 			})
 			return
 		}
 
-		const enabledButtons = createRankedButtons(recruitmentId, false)
+		const enabledButtons = createButtons(recruitmentId, false)
 		await interaction.editReply({
 			embeds: [embed],
 			components: [enabledButtons],
 		})
 	} catch (error) {
-		logger.error('ランク戦募集作成エラー:', error)
+		logger.error('募集作成エラー:', error)
 		await interaction.editReply({
-			content: 'ランク戦募集の作成に失敗しました。',
+			content: '募集の作成に失敗しました。',
 			embeds: [],
 			components: [],
 		})
