+++ src/commands/queue/rank/index.ts
@@ -1,15 +1,16 @@
 import type { ChatInputCommandInteraction } from 'discord.js'
 import { logger } from '@/lib/logger'
 import { apiClient } from '@/utils/api-client'
-import { CAPACITY, createButtons, createEmbed } from './shared'
+import { CAPACITY } from '../shared/constants'
+import { createRankedButtons, createRankedEmbed } from '../shared/embeds'
 
-export const executeCreate = async (interaction: ChatInputCommandInteraction, guildId: string) => {
+export const executeRank = async (interaction: ChatInputCommandInteraction, guildId: string) => {
 	const description = interaction.options.getString('description')
 	const startTime = interaction.options.getString('start_time')
 	const recruitmentId = crypto.randomUUID()
 
-	const embed = createEmbed(false, [], CAPACITY, interaction.user.id, startTime, description)
-	const disabledButtons = createButtons(recruitmentId, true)
+	const embed = createRankedEmbed([], CAPACITY, interaction.user.id, startTime, description)
+	const disabledButtons = createRankedButtons(recruitmentId, true)
 
 	await interaction.reply({
 		embeds: [embed],
@@ -26,30 +27,31 @@ export const executeCreate = async (interaction: ChatInputCommandInteraction, gu
 				channelId: interaction.channelId,
 				messageId: reply.id,
 				creatorId: interaction.user.id,
+				type: 'ranked',
 				anonymous: false,
 				startTime: startTime || undefined,
 			},
 		})
 
 		if (!response.ok) {
-			logger.error('募集作成失敗:', response.status)
+			logger.error('ランク戦募集作成失敗:', response.status)
 			await interaction.editReply({
-				content: '募集の作成に失敗しました。',
+				content: 'ランク戦募集の作成に失敗しました。',
 				embeds: [],
 				components: [],
 			})
 			return
 		}
 
-		const enabledButtons = createButtons(recruitmentId, false)
+		const enabledButtons = createRankedButtons(recruitmentId, false)
 		await interaction.editReply({
 			embeds: [embed],
 			components: [enabledButtons],
 		})
 	} catch (error) {
-		logger.error('募集作成エラー:', error)
+		logger.error('ランク戦募集作成エラー:', error)
 		await interaction.editReply({
-			content: '募集の作成に失敗しました。',
+			content: 'ランク戦募集の作成に失敗しました。',
 			embeds: [],
 			components: [],
 		})
