+++ src/commands/queue/shared/constants.ts
@@ -1,0 +1,13 @@
+import type { LolRole } from '@/constants'
+
+export const CAPACITY = 10
+
+export const INITIAL_RATING = 1500
+
+export const ROLE_LABELS: Record<LolRole, string> = {
+	TOP: 'トップ',
+	JUNGLE: 'ジャングル',
+	MIDDLE: 'ミッド',
+	BOTTOM: 'ボット',
+	SUPPORT: 'サポート',
+}
