+++ src/commands/queue/shared/embeds.ts
@@ -8,37 +8,9 @@ import {
 } from 'discord.js'
 import { COLORS, ROLE_EMOJI } from '@/config'
 import { LOL_ROLES, type LolRole } from '@/constants'
-
-export const CAPACITY = 10
-
-export const ROLE_LABELS: Record<LolRole, string> = {
-	TOP: 'トップ',
-	JUNGLE: 'ジャングル',
-	MIDDLE: 'ミッド',
-	BOTTOM: 'ボット',
-	SUPPORT: 'サポート',
-}
-
-export type Participant = {
-	discordId: string
-	mainRole?: LolRole | null
-	subRole?: LolRole | null
-}
-
-export const parseCustomId = (customId: string) => {
-	const parts = customId.split(':')
-	return {
-		command: parts[0],
-		action: parts[1],
-		recruitmentId: parts[2],
-		originalMessageId: parts[3],
-	}
-}
-
-export const formatRole = (role: LolRole | null | undefined): string => {
-	if (!role) return '-'
-	return `${ROLE_EMOJI[role]} ${ROLE_LABELS[role]}`
-}
+import { CAPACITY, ROLE_LABELS } from './constants'
+import type { Participant, TeamAssignments } from './types'
+import { formatRole } from './utils'
 
 export const createEmbed = (
 	anonymous: boolean,
@@ -178,7 +150,6 @@ export const createClosedEmbed = (
 	return embed
 }
 
-// ランク戦用Embed
 export const createRankedEmbed = (
 	participants: Participant[],
 	capacity: number,
@@ -222,7 +193,6 @@ export const createRankedEmbed = (
 	return embed
 }
 
-// ランク戦用ボタン
 export const createRankedButtons = (recruitmentId: string, disabled: boolean) => {
 	return new ActionRowBuilder<ButtonBuilder>().addComponents(
 		new ButtonBuilder()
@@ -248,7 +218,6 @@ export const createRankedButtons = (recruitmentId: string, disabled: boolean) =>
 	)
 }
 
-// ロール選択セレクトメニュー
 export const createRoleSelectMenu = (recruitmentId: string, type: 'main' | 'sub') => {
 	const options = LOL_ROLES.map((role) =>
 		new StringSelectMenuOptionBuilder()
@@ -266,23 +235,6 @@ export const createRoleSelectMenu = (recruitmentId: string, type: 'main' | 'sub'
 	)
 }
 
-// チーム振り分け用の型
-export type TeamAssignment = {
-	team: 'blue' | 'red'
-	role: LolRole
-	rating: number
-}
-
-export type TeamAssignments = Record<string, TeamAssignment>
-
-export type RatingInfo = {
-	discordId: string
-	rating: number
-	rank: string | null
-	isPlacement: boolean | null
-}
-
-// チーム振り分け結果表示用Embed
 export const createMatchEmbed = (
 	teamAssignments: TeamAssignments,
 	blueVotes: number,
@@ -339,7 +291,6 @@ export const createMatchEmbed = (
 	return embed
 }
 
-// 投票ボタン
 export const createVoteButtons = (matchId: string, disabled = false) => {
 	return new ActionRowBuilder<ButtonBuilder>().addComponents(
 		new ButtonBuilder()
@@ -360,7 +311,6 @@ export const createVoteButtons = (matchId: string, disabled = false) => {
 	)
 }
 
-// 結果確定後のEmbed
 export const createMatchResultEmbed = (
 	winningTeam: 'blue' | 'red',
 	ratingChanges: Array<{
@@ -398,75 +348,4 @@ export const createMatchResultEmbed = (
 		)
 
 	return embed
-}
-
-// Eloに基づくチームバランス関数
-export const balanceTeamsByElo = (
-	participants: Array<{ discordId: string; mainRole?: LolRole | null; subRole?: LolRole | null; rating: number }>,
-): TeamAssignments => {
-	// レーティング順でソート
-	const sorted = [...participants].sort((a, b) => b.rating - a.rating)
-
-	// スネークドラフト方式で分配
-	const blueTeam: typeof sorted = []
-	const redTeam: typeof sorted = []
-
-	sorted.forEach((p, i) => {
-		// 0,3,4,7,8 → Blue, 1,2,5,6,9 → Red (スネーク)
-		const round = Math.floor(i / 2)
-		const isBlue = round % 2 === 0 ? i % 2 === 0 : i % 2 !== 0
-		if (isBlue) {
-			blueTeam.push(p)
-		} else {
-			redTeam.push(p)
-		}
-	})
-
-	// ロールを割り当て
-	const assignRoles = (team: typeof sorted): Array<{ discordId: string; role: LolRole; rating: number }> => {
-		const assigned: Array<{ discordId: string; role: LolRole; rating: number }> = []
-		const usedRoles = new Set<LolRole>()
-
-		// まずメインロールで割り当て
-		for (const p of team) {
-			if (p.mainRole && !usedRoles.has(p.mainRole)) {
-				assigned.push({ discordId: p.discordId, role: p.mainRole, rating: p.rating })
-				usedRoles.add(p.mainRole)
-			}
-		}
-
-		// サブロールで割り当て
-		for (const p of team) {
-			if (!assigned.find((a) => a.discordId === p.discordId)) {
-				if (p.subRole && !usedRoles.has(p.subRole)) {
-					assigned.push({ discordId: p.discordId, role: p.subRole, rating: p.rating })
-					usedRoles.add(p.subRole)
-				}
-			}
-		}
-
-		// 残りは空いているロールを割り当て
-		const remainingRoles = LOL_ROLES.filter((r) => !usedRoles.has(r))
-		for (const p of team) {
-			if (!assigned.find((a) => a.discordId === p.discordId)) {
-				const role = remainingRoles.shift() || 'MIDDLE'
-				assigned.push({ discordId: p.discordId, role, rating: p.rating })
-			}
-		}
-
-		return assigned
-	}
-
-	const blueAssigned = assignRoles(blueTeam)
-	const redAssigned = assignRoles(redTeam)
-
-	const result: TeamAssignments = {}
-	for (const p of blueAssigned) {
-		result[p.discordId] = { team: 'blue', role: p.role, rating: p.rating }
-	}
-	for (const p of redAssigned) {
-		result[p.discordId] = { team: 'red', role: p.role, rating: p.rating }
-	}
-
-	return result
 }
