+++ src/commands/queue/shared/types.ts
@@ -1,0 +1,22 @@
+import type { LolRole } from '@/constants'
+
+export type Participant = {
+	discordId: string
+	mainRole?: LolRole | null
+	subRole?: LolRole | null
+}
+
+export type TeamAssignment = {
+	team: 'blue' | 'red'
+	role: LolRole
+	rating: number
+}
+
+export type TeamAssignments = Record<string, TeamAssignment>
+
+export type RatingInfo = {
+	discordId: string
+	rating: number
+	rank: string | null
+	isPlacement: boolean | null
+}
